+++ backend/rush/src/main/java/rush/rush/controller/ManageGroupController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import rush.rush.dto.CreateGroupRequest;
 import rush.rush.security.CurrentUser;
 import rush.rush.security.user.UserPrincipal;
+import rush.rush.service.CreateGroupService;
 import rush.rush.service.GroupService;
 
 @RequiredArgsConstructor
@@ -19,10 +20,11 @@ import rush.rush.service.GroupService;
 public class ManageGroupController {
 
     private final GroupService groupService;
+    private final CreateGroupService createGroupService;
 
     @PostMapping
     public ResponseEntity<Void> create(@RequestBody CreateGroupRequest createGroupRequest, @CurrentUser UserPrincipal userPrincipal) {
-        Long groupId = groupService.createGroup(createGroupRequest, userPrincipal.getUser());
+        Long groupId = createGroupService.createGroup(createGroupRequest, userPrincipal.getUser());
 
         return ResponseEntity.created(URI.create("/api/groups/" + groupId))
                 .build();
