+++ backend/rush/src/main/java/rush/rush/service/GroupService.java
@@ -9,14 +9,12 @@ import org.springframework.transaction.annotation.Transactional;
 import rush.rush.domain.Group;
 import rush.rush.domain.User;
 import rush.rush.domain.UserGroup;
-import rush.rush.dto.CreateGroupRequest;
 import rush.rush.dto.GroupResponse;
 import rush.rush.dto.GroupSummaryResponse;
 import rush.rush.dto.SimpleUserResponse;
 import rush.rush.repository.GroupRepository;
 import rush.rush.repository.UserGroupRepository;
 import rush.rush.repository.UserRepository;
-import rush.rush.utils.RandomStringGenerator;
 
 @Service
 @RequiredArgsConstructor
@@ -25,18 +23,6 @@ public class GroupService {
     private final GroupRepository groupRepository;
     private final UserGroupRepository userGroupRepository;
     private final UserRepository userRepository;
-
-    @Transactional
-    public Long createGroup(CreateGroupRequest createGroupRequest, User creator) {
-        Group savedGroup = saveGroup(createGroupRequest.getName());
-
-        String invitationCode = generateInvitationCode(savedGroup.getId());
-        savedGroup.setInvitationCode(invitationCode);
-
-        saveUserGroup(savedGroup, creator);
-
-        return savedGroup.getId();
-    }
 
     @Transactional
     public Long join(String invitationCode, User user) {
@@ -79,17 +65,6 @@ public class GroupService {
             .map(member -> new SimpleUserResponse(
                 member.getId(), member.getNickName(), member.getImageUrl()))
             .collect(Collectors.toUnmodifiableList());
-    }
-
-    private Group saveGroup(String groupName) {
-        Group group = Group.builder()
-                .name(groupName)
-                .build();
-        return groupRepository.save(group);
-    }
-
-    private String generateInvitationCode(Long groupId) {
-        return RandomStringGenerator.generate(5) + groupId;
     }
 
     private boolean hasJoined(Long groupId, Long userId) {
